Support Windows DinD-like
