+++ main.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/rancher/dapper/file"
 	"github.com/sirupsen/logrus"
 	"github.com/urfave/cli.v1"
-	"github.com/rancher/dapper/file"
 )
 
 var (
